CRUD; Read Update
In CRUD,  I added a code in Node, that allows you to read new comments. 
I installed the UUID package using NPM. The UUID package creates a Unique User ID-entification for each comments created. 
That way, each comment can be easily accessed (for viewing, updating, and/or deletion) when needed using this its unique ID. 

Now, that leads us to the "U" in CRUD, which stands for "Update" :)